requirements.txt
asgiref==3.7.1
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pymongo==4.3.3
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.1
